requirements.txt
llama-cpp-python
phonemizer
inflect
unidecode
pygame
scipy
BeautifulSoup4
nltk==3.8.1
wikipedia
pvporcupine
pvrecorder
vosk
sounddevice
colorama
geotext
urllib
geotext
timezonefinder
requests
dateparser
timefhuman
google-api-python-client
pafy
python-vlc
yt-dlp